requirements.txt
ipykernel
pandas
pyspark
numpy
pyarrow
sqlalchemy
psycopg2-binary
geopy
certifi
matplotlib
seaborn
streamlit
plotly
python-dotenv